--- app/src/androidTest/java/net/ommoks/azza/android/app/passonnotifications/ExampleInstrumentedTest.kt
@@ -1,24 +1,0 @@
-package net.ommoks.azza.android.app.passonnotifications
-
-import androidx.test.platform.app.InstrumentationRegistry
-import androidx.test.ext.junit.runners.AndroidJUnit4
-
-import org.junit.Test
-import org.junit.runner.RunWith
-
-import org.junit.Assert.*
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("net.ommoks.azza.android.app.passonnotifications", appContext.packageName)
-    }
-}
