+++ app/src/main/java/net/ommoks/azza/android/app/pass_on_messages/PomApplication.kt
@@ -1,0 +1,8 @@
+package net.ommoks.azza.android.app.pass_on_messages
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class PomApplication : Application() {
+}
