+++ app/src/main/java/net/ommoks/azza/android/app/pass_on_messages/common/Constants.kt
@@ -1,0 +1,7 @@
+package net.ommoks.azza.android.app.pass_on_messages.common
+
+object Constants {
+
+    const val FILTER_FILE = "filter.json"
+    const val LOG_FILE = "log.txt"
+}
