--- app/src/main/java/net/ommoks/azza/android/app/passonnotifications/PonApplication.kt
@@ -1,8 +1,0 @@
-package net.ommoks.azza.android.app.passonnotifications
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class PonApplication : Application() {
-}
