--- app/src/main/java/net/ommoks/azza/android/app/passonnotifications/common/Constants.kt
@@ -1,7 +1,0 @@
-package net.ommoks.azza.android.app.passonnotifications.common
-
-object Constants {
-
-    const val FILTER_FILE = "filter.json"
-    const val LOG_FILE = "log.txt"
-}
