--- app/src/main/java/net/ommoks/azza/android/app/passonnotifications/data/MainRepository.kt
@@ -1,8 +1,0 @@
-package net.ommoks.azza.android.app.passonnotifications.data
-
-import net.ommoks.azza.android.app.passonnotifications.data.model.Filter
-
-interface MainRepository {
-    suspend fun saveFilters(filters: List<Filter>)
-    suspend fun loadFilters() : List<Filter>
-}
