--- app/src/main/java/net/ommoks/azza/android/app/passonnotifications/data/di/RepositoryModule.kt
@@ -1,24 +1,0 @@
-package net.ommoks.azza.android.app.passonnotifications.data.di
-
-import android.content.Context
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import net.ommoks.azza.android.app.passonnotifications.data.MainRepository
-import net.ommoks.azza.android.app.passonnotifications.data.datasource.FileDataSource
-import net.ommoks.azza.android.app.passonnotifications.data.impl.MainRepositoryImpl
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-class RepositoryModule {
-
-    @Singleton
-    @Provides
-    fun provideMainRepository(
-        @ApplicationContext appContext: Context,
-        fileDataSource: FileDataSource
-    ) : MainRepository = MainRepositoryImpl(appContext, fileDataSource)
-}
