+++ app/src/main/res/layout/activity_main.xml
@@ -24,7 +24,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container"
-        android:name="net.ommoks.azza.android.app.passonnotifications.ui.MainFragment"
+        android:name="net.ommoks.azza.android.app.pass_on_messages.ui.MainFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         tools:layout="@layout/fragment_main"
